+++ pluribus/util.hpp
@@ -1,10 +1,10 @@
 #pragma once
 
-#include <string>
-#include <fstream>
 #include <filesystem>
-#include <json/json.hpp>
+#include <fstream>
+#include <string>
 #include <hand_isomorphism/hand_index.h>
+#include <json/json.hpp>
 
 namespace pluribus {
 
@@ -17,7 +17,7 @@ uint8_t card_to_idx(const std::string& card);
 std::string idx_to_card(int idx);
 void str_to_cards(const std::string &card_str, uint8_t cards[]);
 std::vector<uint8_t> str_to_cards(const std::string &card_str);
-std::string cards_to_str(const uint8_t cards[], int n);
+std::string cards_to_str(const uint8_t cards[], size_t n);
 std::string cards_to_str(const std::vector<uint8_t>& cards);
 int n_board_cards(int round);
 int init_indexer(hand_indexer_t& indexer, int round);
